+++ backend/document-core/src/main/java/com/docflow/documentcore/domain/model/PermisoDocumentoRol.java
@@ -7,8 +7,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hibernate.annotations.Filter;
-import org.hibernate.annotations.FilterDef;
-import org.hibernate.annotations.ParamDef;
 
 import java.time.OffsetDateTime;
 
@@ -41,10 +39,10 @@ public class PermisoDocumentoRol {
     private Long documentoId;
     
     @Column(name = "rol_id", nullable = false)
-    private Integer rolId;
+    private Long rolId;
     
     @Column(name = "organizacion_id", nullable = false)
-    private Integer organizacionId;
+    private Long organizacionId;
     
     @Enumerated(EnumType.STRING)
     @Column(name = "nivel_acceso", nullable = false, length = 20)
